requirements.txt
streamlit==1.35.0
plotly==5.22.0
pandas==2.2.2
openpyxl==3.1.2
google-auth==2.29.0
google-api-python-client==2.128.0
